Add Matrix class
Matrix represents a sparse or dense matrix of 1 or 2 dimensions, with
possibly unknown shape and dtype. This comes up frequently when applying
operations to `db.Bag` objects that result in a tall skinny matrix,
where the height (and potentially width) are unknown. For most machine
learning applications the height is unneeded, and everything works out
fine.

The class implements almost no operations - all that's really needed is
a `map_partitions` function to map a function across all blocks, and
a finalizer to ensure that the output class is correct (sparse ->
sparse, dense -> dense, concatenating as appropriate).