Chained now masquerades as the wrapped estimator
Previously `Chained` acted as a meta-estimator, meaning that
`est.get_params() != Chained(est).get_params()`. This was nice, because
there may be a time in the future when meta-estimators like `Chained`
have non-estimator parameters (e.g. a `n_iter` parameter for Chained).
However, this creates a dissonance when `Chained(est).compute() -> est`,
as the attributes on the computed object don't match the attributes on
the proxy object.

This PR removes that aspect, so that Chained and its contained estimator
both act the same.