A few tweaks (#47)
* No longer accept address for `scheduler` kwarg

This was causing things to lockup if an invalid parameter was passed in.
We now accept a `Client` instead, which is both simpler and more robust.

* Squash a warning in feature_union test

We can't turn off this warning since it's from inside joblib, ignore it
for now.

* Use catch_warnings instead

`ignore_warnings` doesn't work with parametrize in python 2.7 due to how
parametrize inspects test function signatures. Use the `catch_warnings`
contextmanager instead.