Call map_blocks with dtype in ParallelPostFit.fit
Instead of letting `map_blocks` guess the `dtype` needed by `transform`,
provide it the `dtype` of `X`. This may not always be right, but is more
likely right than wrong. Further it gives the user a mechanism to fix
the `dtype` returned by `map_blocks` when it's heuristics for detection
fail, which is an improvement over the current behavior that does not
supply such an option.