Add basic incremental training function
This implements a simple approach to do hyper-parameter optimization on
incrementally trained models.  It is heavily insprired by (and steals content
from) #221 but attempts less.

Currently it implements a function ``fit``, which takes in a model, training
and testing data and a parameter space.  It creates many models and trains them
on a block of data, scores them, and then keeps only the better performing
ones.  It continues to do this until only one is selected.

This function is not intended for public users, but is hopefully something we
can play with as we eventually build a full solution.  It is decently
configurable today and we will need to find sane defaults.  Notably, this
function does not take opinions on the following:

1.  How testing data should be separated from training data,
    currently they are both explicitly provided
2.  How many parameters to start with, or how to decrease this over time
    These are given as inputs with an integer and function respectively

Additionally it does not, but probably should provide configurations for the
following:

1.  How to select models from the current group
2.  How to decide on stopping criteria
3.  How any ability actively decide the number of surviving models
    based on the scores received so far

I believe that this implementation is a decent start though and should be a
conversation piece.  I think that it makes some decisions that we don't want to
think about yet (online determination of how many models to keep) while giving
us knobs for other parameters that I think we'll want to play with (model decay
functions, how to handle testing, etc..)